backend revision second day done (mapping corrected for frontEnd and backend is in final state at least for now)

+++ src/main/java/com/example/SpringLogin/Entities/Question.java
@@ -22,7 +22,7 @@ public class Question implements Serializable {
     private String content;
     @Column(nullable = false)
     private String typeAnswer;
-    private int points;
+    private float points;
     private String tags;
     @Column(nullable = false)
     private Timestamp dateCreation;
